Make the jemalloc dependency optional
