+++ modules/malloc/CMakeLists.txt
@@ -3,12 +3,13 @@
 file(GLOB_RECURSE BASIC_SRC_FILES "${CMAKE_CURRENT_SOURCE_DIR}" "*.cc")
 
 add_library(vineyard_malloc ${BASIC_SRC_FILES})
-target_link_libraries(vineyard_malloc vineyard_client
-                                      ${ARROW_SHARED_LIB}
+target_link_libraries(vineyard_malloc PUBLIC vineyard_client
+                                             ${ARROW_SHARED_LIB}
 )
 
-if(VINEYARD_MOD_SRCS)
-    add_dependencies(vineyard_malloc vineyard_malloc)
+if(WITH_ALLOCATOR STREQUAL "jemalloc")
+    target_link_libraries(vineyard_malloc PRIVATE jemalloc ${CMAKE_DL_LIBS})
+    target_compile_options(vineyard_malloc PRIVATE -DWITH_JEMALLOC)
 endif()
 
 install_vineyard_target(vineyard_malloc)
